EasyS7/requirements.txt
python-snap7==0.11
